requirements.txt
iniconfig==2.3.0
packaging==26.1
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
PyVISA==1.16.2
PyYAML==6.0.3
typing_extensions==4.15.0
